Refactor redundant register_watcher[s] to Connection
