+++ cassandra/io/asyncorereactor.py
@@ -40,7 +40,6 @@ except ImportError:
 from cassandra.connection import (Connection, ConnectionShutdown,
                                   ConnectionException, NONBLOCKING,
                                   Timer, TimerManager)
-from cassandra.protocol import RegisterMessage
 
 log = logging.getLogger(__name__)
 
@@ -318,14 +317,3 @@ class AsyncoreConnection(Connection, asyncore.dispatcher):
 
     def readable(self):
         return self._readable or (self.is_control_connection and not (self.is_defunct or self.is_closed))
-
-    def register_watcher(self, event_type, callback, register_timeout=None):
-        self._push_watchers[event_type].add(callback)
-        self.wait_for_response(
-            RegisterMessage(event_list=[event_type]), timeout=register_timeout)
-
-    def register_watchers(self, type_callback_dict, register_timeout=None):
-        for event_type, callback in type_callback_dict.items():
-            self._push_watchers[event_type].add(callback)
-        self.wait_for_response(
-            RegisterMessage(event_list=type_callback_dict.keys()), timeout=register_timeout)
